+++ src/main/java/info/faceland/strife/data/buff/LoadedBuff.java
@@ -6,11 +6,11 @@ import java.util.Map;
 public class LoadedBuff {
 
   private final String name;
-  private final Map<StrifeStat, Double> stats;
+  private final Map<StrifeStat, Float> stats;
   private final int maxStacks;
-  private final double seconds;
+  private final float seconds;
 
-  public LoadedBuff(String name, Map<StrifeStat, Double> stats, int maxStacks, double seconds) {
+  public LoadedBuff(String name, Map<StrifeStat, Float> stats, int maxStacks, float seconds) {
     this.name = name;
     this.stats = stats;
     this.maxStacks = maxStacks;
@@ -21,7 +21,7 @@ public class LoadedBuff {
     return name;
   }
 
-  public Map<StrifeStat, Double> getStats() {
+  public Map<StrifeStat, Float> getStats() {
     return stats;
   }
