+++ src/main/java/info/faceland/strife/data/champion/PlayerEquipmentCache.java
@@ -18,12 +18,12 @@ public class PlayerEquipmentCache {
 
   private final Map<EquipmentSlot, Integer> slotHashCodeMap = new HashMap<>();
 
-  private final Map<EquipmentSlot, Map<StrifeStat, Double>> slotStatMap = new HashMap<>();
+  private final Map<EquipmentSlot, Map<StrifeStat, Float>> slotStatMap = new HashMap<>();
   private final Map<EquipmentSlot, List<LoreAbility>> slotAbilityMap = new HashMap<>();
   private final Map<EquipmentSlot, List<StrifeTrait>> slotTraitMap = new HashMap<>();
 
   private final Map<TriggerType, Set<LoreAbility>> loreAbilities = new HashMap<>();
-  private final Map<StrifeStat, Double> combinedStats = new HashMap<>();
+  private final Map<StrifeStat, Float> combinedStats = new HashMap<>();
   private final Set<StrifeTrait> combinedTraits = new HashSet<>();
 
   public final static EquipmentSlot[] itemSlots = EquipmentSlot.values();
@@ -40,12 +40,12 @@ public class PlayerEquipmentCache {
     }
   }
 
-  public void setSlotStats(EquipmentSlot slot, Map<StrifeStat, Double> stats) {
+  public void setSlotStats(EquipmentSlot slot, Map<StrifeStat, Float> stats) {
     this.slotStatMap.get(slot).clear();
     this.slotStatMap.get(slot).putAll(stats);
   }
 
-  public Map<StrifeStat, Double> getSlotStats(EquipmentSlot slot) {
+  public Map<StrifeStat, Float> getSlotStats(EquipmentSlot slot) {
     return this.slotStatMap.get(slot);
   }
 
@@ -81,7 +81,7 @@ public class PlayerEquipmentCache {
     this.slotTraitMap.get(slot).clear();
   }
 
-  public Map<StrifeStat, Double> getCombinedStats() {
+  public Map<StrifeStat, Float> getCombinedStats() {
     return combinedStats;
   }
